moved hooks and decorator
