+++ frontend/app/src/entities/diff/ui/diff-tree.tsx
@@ -2,7 +2,7 @@ import { DiffNode } from "@/entities/diff/node-diff/types";
 import { DiffBadge } from "@/entities/diff/node-diff/utils";
 import { TREE_ROOT_ID } from "@/entities/ipam/constants";
 import { EMPTY_TREE, addItemsToTree } from "@/entities/ipam/ipam-tree/utils";
-import { useSchema } from "@/entities/schema/hooks/useSchema";
+import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import { Tooltip } from "@/shared/components/ui/tooltip";
 import { Tree, TreeItemProps, TreeProps } from "@/shared/components/ui/tree";
 import { Icon } from "@iconify-icon/react";
