+++ frontend/app/src/entities/nodes/object/ui/object-table/cells/object-actions-cell.tsx
@@ -2,7 +2,7 @@ import ObjectItemEditComponent from "@/entities/nodes/object-item-edit/object-it
 import { DeleteObjectModal } from "@/entities/nodes/object/ui/delete-object-modal";
 import { getObjectDetailsUrl2 } from "@/entities/nodes/utils";
 import { Permission } from "@/entities/permission/types";
-import { useSchema } from "@/entities/schema/hooks/useSchema";
+import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import { queryClient } from "@/shared/api/rest/client";
 import { Button } from "@/shared/components/buttons/button-primitive";
 import SlideOver, { SlideOverTitle } from "@/shared/components/display/slide-over";
