+++ frontend/app/src/entities/role-manager/ui/object-permissions-form.tsx
@@ -18,7 +18,7 @@ import { useAtomValue } from "jotai";
 import { FieldValues, useForm } from "react-hook-form";
 import { toast } from "react-toastify";
 
-import { useSchema } from "@/entities/schema/hooks/useSchema";
+import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import DropdownField from "@/shared/components/form/fields/dropdown.field";
 import RelationshipManyField from "@/shared/components/form/fields/relationship-many.field";
 import { NameSelect } from "@/shared/components/form/name-select";
