+++ frontend/app/src/entities/role-manager/ui/roles.tsx
@@ -8,7 +8,7 @@ import { Pagination } from "@/shared/components/ui/pagination";
 import { useAtomValue } from "jotai";
 import { useState } from "react";
 
-import { useSchema } from "@/entities/schema/hooks/useSchema";
+import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import graphqlClient from "@/shared/api/graphql/graphqlClientApollo";
 import useQuery from "@/shared/api/graphql/useQuery";
 import { Button } from "@/shared/components/buttons/button-primitive";
