+++ frontend/app/src/entities/schema/ui/hooks/useSchema.ts
