+++ frontend/app/src/shared/components/layout/breadcrumb-navigation/items/breadcrumb-schema-selector.tsx
@@ -1,6 +1,6 @@
 import { PROFILE_KIND } from "@/config/constants";
 import { getObjectDetailsUrl2 } from "@/entities/nodes/utils";
-import { useSchema } from "@/entities/schema/hooks/useSchema";
+import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import { BreadcrumbLink } from "@/shared/components/layout/breadcrumb-navigation/items/breadcrumb-link";
 import BreadcrumbLoading from "@/shared/components/layout/breadcrumb-navigation/items/breadcrumb-loading";
 import { breadcrumbActiveStyle } from "@/shared/components/layout/breadcrumb-navigation/style";
