+++ frontend/app/tests/integrations/screens/artifact-diff.cy.tsx
@@ -7,8 +7,8 @@ import { ACCESS_TOKEN_KEY } from "../../../src/config/localStorage";
 import { AuthProvider } from "../../../src/entities/authentication/ui/useAuth";
 import { ArtifactsDiff } from "../../../src/entities/diff/artifact-diff/artifacts-diff";
 import { proposedChangedState } from "../../../src/entities/proposed-changes/stores/proposedChanges.atom";
-import { withSchemaContext } from "../../../src/entities/schema/decorators/withSchemaContext";
 import { nodeSchemasAtom } from "../../../src/entities/schema/stores/schema.atom";
+import { withSchemaContext } from "../../../src/entities/schema/ui/decorators/withSchemaContext";
 import { encodeJwt } from "../../../src/shared/utils/common";
 import { accountDetailsMocksSchema } from "../../mocks/data/account";
 import {
